Automatically backed up by Learn

+++ db/migrate/001_create_costumes.rb
@@ -9,15 +9,14 @@
 # value types according to the spec
 
 class CreateCostumes < ActiveRecord::Migration[4.2]
-  
-  def change 
+  def change
     create_table :costumes do |t|
-      t.string :name 
-      t.string :size 
-      t.string :image_url 
-      t.integer :price 
-      
+      t.string :name
+      t.string :size
+      t.string :image_url
+      t.float :price
+
       t.timestamps null: false
-    end 
-  end 
-end 
+    end
+  end
+end
